requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
pypdf
python-docx
python-multipart
google-genai
python-dotenv
streamlit
fpdf2
